Add a section on cookie representation
This defines all the cookie fields we discussed. We will need to fiddle with the initial values some more as we go and also with some of the types potentially.

Ideally we'd reference the Infra Standard directly for all of the definitions, but I'm not sure how to do that.